# Performance improvement by moving .to(device)