@alejandromarcu 2025-04-24
Also, it's interesting to compare the same run in main and in this PR.
Notice that reward and loss are overlapping for both runs, so it gives me more confidence that the functionality hasn't changed.
Then, you can see that with the optimization, the run is much shorter but the GPU runs hotter, probably because we're making it do the same work without pauses to move data around. 


<img width="1660" alt="image" src="https://github.com/user-attachments/assets/1ef314b0-0856-42ec-8fb3-40e68d4966ad" />