@adamantivm 2025-04-24
Nice finding! Let me see if I can try it and compare. I also want to see if any of this translates to the SB3 version of training or playing.