@adamantivm 2025-04-24
Run without optimization:
```
Training completed!
         22901500 function calls (22078465 primitive calls) in 136.779 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1   63.012   63.012   63.012   63.012 serialization.py:1120(_save)
    41330   44.478    0.001   44.478    0.001 {method 'item' of 'torch._C.TensorBase' objects}
   765850    5.769    0.000    5.784    0.000 {method 'to' of 'torch._C.TensorBase' objects}
    99/89    2.770    0.028    2.773    0.031 {method 'recv' of '_socket.socket' objects}
     1968    2.569    0.001    8.310    0.004 trainable_agent.py:379(_prepare_batch)
       74    1.597    0.022    1.597    0.022 {method 'poll' of 'select.epoll' objects}
     1974    1.569    0.001    1.569    0.001 {built-in method torch._C._cuda_getDeviceCount}
       10    1.424    0.142    1.424    0.142 {method 'read' of '_ssl._SSLSocket' objects}
  2250375    1.282    0.000    2.171    0.000 quoridor_env.py:166(is_wall_between)
```