@adamantivm 2025-04-24
I lost the run with optimization, becuase my computer crashed, but it looked almost the same.